Fix decoding of JBIG2 integers for values with large magnitude.
The JBIG2 spec says that 32 bits are used to encode (val - 4436)
and does not define limits for the val, so it could be up to
(2^32 + 4435). On the negative side, 32 bits are used to encode
(-val - 4436), making -(2^32 + 4436) the lowest value
representable.

By using a 32-bit integer to store the magnitude, we only supported
up to 2^31 - 1. Larger values became negative. The lowest value we
decoded correctly was -(2^31 - 1).

The typing changes should accommodate these new values, though the
method signature still requires casting down and losing the range
above 2^31 - 1 and below -2^31.

Bug: chromium:761666

+++ core/fxcodec/jbig2/JBig2_ArithIntDecoder.cpp
@@ -49,6 +49,9 @@ CJBig2_ArithIntDecoder::~CJBig2_ArithIntDecoder() {}
 
 bool CJBig2_ArithIntDecoder::decode(CJBig2_ArithDecoder* pArithDecoder,
                                     int* nResult) {
+  // This decoding algorithm is explained in "Annex A - Arithmetic Integer
+  // Decoding Procedure" on page 113 of the JBIG2 specification (ISO/IEC FCD
+  // 14492).
   int PREV = 1;
   const int S = pArithDecoder->DECODE(&m_IAx[PREV]);
   PREV = ShiftOr(PREV, S);
@@ -56,7 +59,7 @@ bool CJBig2_ArithIntDecoder::decode(CJBig2_ArithDecoder* pArithDecoder,
   const size_t nDecodeDataIndex =
       RecursiveDecode(pArithDecoder, &m_IAx, &PREV, 0);
 
-  int nTemp = 0;
+  uint32_t nTemp = 0;
   for (int i = 0; i < g_ArithIntDecodeData[nDecodeDataIndex].nNeedBits; ++i) {
     int D = pArithDecoder->DECODE(&m_IAx[PREV]);
     PREV = ShiftOr(PREV, D);
@@ -64,12 +67,12 @@ bool CJBig2_ArithIntDecoder::decode(CJBig2_ArithDecoder* pArithDecoder,
       PREV = (PREV & 511) | 256;
     nTemp = ShiftOr(nTemp, D);
   }
-  int nValue = g_ArithIntDecodeData[nDecodeDataIndex].nValue;
+  int64_t nValue = g_ArithIntDecodeData[nDecodeDataIndex].nValue;
   nValue += nTemp;
   if (S == 1 && nValue > 0)
     nValue = -nValue;
 
-  *nResult = nValue;
+  *nResult = static_cast<int>(nValue);
   return S != 1 || nValue != 0;
 }
